split API,remake api.js with ts
