+++ src/pages/repositories/repository/repo-comp/error/error.tsx
@@ -30,7 +30,6 @@ const RepositoryInDeletionContainer = (repoId:string) => {
     );
 };
 
-
 export const RepoError = (error:RepositoryDeletionError| null) => {
     if (error instanceof RepositoryDeletionError) {
         return <RepositoryInDeletionContainer repoId={error.repoId}/>;
